fix: moved components and styles to their correct folders
