+++ src/app/app-routing.module.ts
@@ -3,11 +3,11 @@ import { HomeComponent } from "./pages/home/home.component";
 import { ProfileComponent } from "./pages/profile/profile.component";
 import { ExternalApiComponent } from "./pages/external-api/external-api.component";
 import { ErrorComponent } from "./pages/error/error.component";
-import { ProductManagementComponent } from "./pages/product-management/product-management.component";
+import { ProductManagementComponent } from "./pages/admin/product-management/product-management.component";
 import { AuthGuard } from "../app/core/guards/auth.guard"; // Import your custom AuthGuard
-import { ProductDetailsComponent } from "./components/product-details/product-details.component";
-import { CategoryManagementComponent } from "./components/category-management/category-management.component";
-import { CategoryDetailsComponent } from "./components/category-details/category-details.component";
+import { ProductDetailsComponent } from "./pages/product-details/product-details.component";
+import { CategoryManagementComponent } from "./pages/admin/category-management/category-management.component";
+import { CategoryDetailsComponent } from "./pages/category-details/category-details.component";
 import { BannerManagementComponent } from "./pages/banner-management/banner-management.component";
 import { OrderComponent } from "./pages/order/order.component";
