+++ src/app/app.component.less
@@ -1,4 +1,4 @@
-@import "../app/components/shared/styles/media-query.less";
+@import "./shared/styles/media-query.less";
 
 .main-container {
   display: grid;
