+++ src/app/components/category-list/category-list.component.less
@@ -1,6 +1,6 @@
 @import "../../shared/styles/global_colors.less";
 @import "../../shared/styles/breakpoints.less";
-@import "../shared/styles/media-query.less";
+@import "../../shared/styles/media-query.less";
 
 .category-container {
   display: flex;
