+++ src/app/components/item-list/item-list.component.less
@@ -1,6 +1,6 @@
 @import "../../shared/styles/global_colors.less";
 @import "../../shared/styles/breakpoints.less";
-@import "../shared/styles/media-query.less";
+@import "../../shared/styles/media-query.less";
 
 .product-list {
   display: flex;
