+++ src/app/components/product-list/product-list.component.less
@@ -1,11 +1,10 @@
 @import "../../shared/styles/global_colors.less";
 @import "../../shared/styles/breakpoints.less";
-@import "../shared/styles/media-query.less";
+@import "../../shared/styles/media-query.less";
 .product-list {
   display: flex;
   flex-direction: column;
   align-items: center;
-
 
   h2 {
     margin-bottom: 16px;
