+++ src/app/components/shared/footer/footer.component.less
@@ -1,4 +1,4 @@
-@import "../styles/colors.less";
+@import "../../../shared/styles/colors.less";
 
 .total-footer {
   margin: 0;
