+++ src/app/components/shared/nav-bar/nav-bar.component.less
@@ -1,5 +1,5 @@
-@import "../styles/media-query.less";
-@import "../styles/colors.less";
+@import "../../../shared/styles/media-query.less";
+@import "../../../shared/styles/colors.less";
 @import "../../../shared/styles/top-left-text.less";
 
 .navbar {
