+++ src/app/components/shared/side-bar/side-bar.component.less
@@ -1,4 +1,4 @@
-@import "../styles/colors.less";
+@import "../../../shared/styles/colors.less";
 @import "../../../shared/styles/breakpoints.less";
 
 .sidebar-container {
