+++ src/app/pages/category-details/category-details.component.less
