+++ src/app/pages/category-details/category-details.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnDestroy, OnInit } from "@angular/core";
-import { SideBarComponent } from "../shared/side-bar/side-bar.component";
-import { ProductListComponent } from "../product-list/product-list.component";
-import { HorizontalSliderComponent } from "../horizontal-slider/horizontal-slider.component";
+import { SideBarComponent } from "../../components/shared/side-bar/side-bar.component";
+import { ProductListComponent } from "../../components/product-list/product-list.component";
+import { HorizontalSliderComponent } from "../../components/horizontal-slider/horizontal-slider.component";
 import { ApiService } from "src/app/api.service";
 import { Category } from "src/app/interfaces/category.interface";
 import { CapitalizePipe } from "../../shared/pipes/capitalize.pipe";
 import { FilterService } from "../../core/services/filter.service";
-import { LoadingComponent } from "../shared/loading/loading.component";
+import { LoadingComponent } from "../../components/shared/loading/loading.component";
 
 @Component({
   selector: "app-category-details",
