+++ src/app/pages/home/home.component.less
@@ -1,6 +1,6 @@
 @import "../../shared/styles/global_colors.less";
 @import "../../shared/styles/breakpoints.less";
-@import "../../components/shared/styles/media-query.less";
+@import "../../shared/styles/media-query.less";
 h4 {
   text-align: center;
   color: #7c7c7c;
