+++ src/app/pages/product-details/product-details.component.less
@@ -1,5 +1,5 @@
-@import "../shared/styles/media-query.less";
-@import "../shared/styles/colors.less";
+@import "../../shared/styles/media-query.less";
+@import "../../shared/styles/colors.less";
 
 @main-color: #007b00;
 @text-color: #666;
