+++ src/app/shared/styles/colors.less
